add types folder into tags feature, add user id property into receiveEntitiesAction
